Build: Compile with -Wfloat-equal

+++ Qt/Qt_Fonts.cpp
@@ -337,13 +337,13 @@ void GFX_IncFontSize(struct Tracker_Windows *tvisual, int pixels){
     // not used
     editor->font.setPixelSize(editor->font.pixelSize()+pixels);
   }else{
-    double org_size = editor->font.pointSize();
+    int org_size = editor->font.pointSize();
     for(int i=1;i<100;i++){
       int new_font_size = org_size+(i*pixels);
       if(new_font_size<2)
         return;
       editor->font.setPointSize(new_font_size);
-      if(editor->font.pointSize()!=org_size)
+      if(editor->font.pointSize() != org_size)
         goto exit;
     }
     for(float i=1;i<100;i++){
